직접 풀어보기 4-3 입력 실수값 계산

+++ Mirirm_project_0412_2/app/src/main/java/kr/hs/emirim/s2114/mirirm_project_0412_2/MainActivity.java
@@ -40,9 +40,9 @@ public class MainActivity extends AppCompatActivity {
                         Toast.LENGTH_LONG).show();
                 return;
             }
-            int num1 = Integer.parseInt(edit1Str);
-            int num2 = Integer.parseInt(edit2Str);
-            int result = 0;
+            double num1 = Double.parseDouble(edit1Str);
+            double num2 = Double.parseDouble(edit2Str);
+            double result = 0;
 
             switch (view.getId()) {
                 case R.id.btn_plus:
